@@ -1,6 +1,6 @@
 package com.erikagtierrez.multiple_media_picker;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 /**
  *
