@@ -2,14 +2,6 @@ package com.erikagtierrez.multiple_media_picker;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.TabLayout;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.view.WindowManager;
 
@@ -17,9 +9,18 @@ import com.erikagtierrez.multiple_media_picker.fragment.AudioFragment;
 import com.erikagtierrez.multiple_media_picker.fragment.DocumentFragment;
 import com.erikagtierrez.multiple_media_picker.fragment.ImageFragment;
 import com.erikagtierrez.multiple_media_picker.fragment.VideoFragment;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.tabs.TabLayout;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
+import androidx.viewpager.widget.ViewPager;
 
 public class Gallery extends AppCompatActivity {
     private TabLayout tabLayout;
