@@ -1,7 +1,5 @@
 package com.erikagtierrez.multiple_media_picker.adapter;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
 import android.widget.ImageView;
@@ -13,6 +11,8 @@ import com.erikagtierrez.multiple_media_picker.R;
 import com.erikagtierrez.multiple_media_picker.model.FileItem;
 
 import java.util.ArrayList;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 public class ListAdapter extends RecyclerView.Adapter<ListAdapter.ViewHolder> {
     private ArrayList<FileItem> fileItems;
