@@ -5,10 +5,6 @@ import android.content.Intent;
 import android.database.Cursor;
 import android.os.Bundle;
 import android.provider.MediaStore;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.GestureDetector;
 import android.view.LayoutInflater;
 import android.view.MotionEvent;
@@ -25,7 +21,12 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 
-public class AudioFragment extends Fragment{
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.DefaultItemAnimator;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+public class AudioFragment extends Fragment {
     private static RecyclerView recyclerView;
     private BucketsAdapter mAdapter;
     private final String[] projection = new String[]{ MediaStore.Audio.Media.ALBUM, MediaStore.Audio.Media.DATA };
