@@ -1,14 +1,10 @@
 package com.erikagtierrez.multiple_media_picker.fragment;
 
-import android.app.Fragment;
 import android.content.Context;
 import android.content.Intent;
 import android.database.Cursor;
 import android.os.Bundle;
 import android.provider.MediaStore;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.GestureDetector;
 import android.view.LayoutInflater;
 import android.view.MotionEvent;
@@ -24,7 +20,12 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 
-public class VideoFragment extends Fragment{
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.DefaultItemAnimator;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+public class VideoFragment extends Fragment {
     private static RecyclerView recyclerView;
     private BucketsAdapter mAdapter;
     private List<String> bucketNames= new ArrayList<>();
